Read Metadata for a Collection

Read Metadata for a Collection

db_collection_read_metadata(
  con,
  collection_name,
  collection_owner,
  valid_on = NULL,
  locale = NULL,
  schema = "timeseries"
)

Arguments

con
RPostgres connection object.

collection_name
character name of a collection to read. Collection are bookmark lists that contain time series keys.

collection_owner
character username that is the owner of a collection.

valid_on
character representation of a date in the form of 'YYYY-MM-DD'. valid_on selects the version of a time series that is valid at the specified time.

locale
character indicating the language of the meta information to be store. We recommend to use ISO country codes to represent languages. Defaults to NULL. When local is set to NULL, metadata are stored without localization. Note that, when localizing meta information by assigning a language, multiple meta information objects can be stored for a single time series.

schema
character name of the database schema. Defaults to 'timeseries'

character
collection_owner character name of the collection owner.

Value

See also