Read Time Series Metadata

Read meta information given a vector of time series identifiers.

db_metadata_read(
  con,
  ts_keys,
  valid_on = NULL,
  regex = FALSE,
  locale = NULL,
  schema = "timeseries"
)

Arguments

con

RPostgres connection object.

ts_keys

character vector of time series identifiers.

valid_on

character representation of a date in the form of 'YYYY-MM-DD'. valid_on selects the version of a time series that is valid at the specified time.

regex

boolean indicating if ts_keys should be interpreted as a regular expression pattern. Defaults to FALSE.

locale

character indicating the language of the meta information to be store. We recommend to use ISO country codes to represent languages. Defaults to NULL. When local is set to NULL, metadata are stored without localization. Note that, when localizing meta information by assigning a language, multiple meta information objects can be stored for a single time series.

schema

character name of the database schema. Defaults to 'timeseries'

Value

list of tsmeta objects.

See also