Create an Entry in the Release Calendar

The idea of the release calendar is to set a release date for some time series that might be in the database already but should not be publicly available before a specific date, e.g., a press release. Since publishing is simply a matter of changing the access level, an update of the access levels could be triggered based on the release information in a release table. Only timeseries admins may create and modify releases.

db_release_create(
  con,
  id,
  title,
  release_date,
  datasets,
  target_year = year(release_date),
  target_period = month(release_date),
  target_frequency = 12,
  note = NULL,
  schema = "timeseries"
)

Arguments

con

RPostgres connection object.

id

Identifier for the release e.g. 'gdb_may_2020'

title

Display title for the release

release_date

Timestamp when the release is to occur

datasets

character vector of the datasets. Dataset is a group of time series.

target_year

Year observed in the data

target_period

Period observed in the data (e.g. month, quarter)

target_frequency

Frequency of the data (e.g. 4 for quarterly)

note

Additional remarks about the release.

schema

character name of the database schema. Defaults to 'timeseries'

Value

a status list

Details

target_period changes meaning depending on the frequency of the release. e.g. period 2 for quarterly data (reference_frequency = 4) means Q2 whereas period 2 for monthly data (frequency 12) means February In other words: target_year and target_period mark the end of the time series in the release.

See also