Common parameters

Common parameters

Arguments

con

RPostgres connection object.

schema

character name of the database schema. Defaults to 'timeseries'

ts_keys

character vector of time series identifiers.

dataset

character name of the dataset. Datasets are group of time series.

datasets

character vector of the datasets. Dataset is a group of time series.

valid_on

character representation of a date in the form of 'YYYY-MM-DD'. valid_on selects the version of a time series that is valid at the specified time.

valid_from

character representation of a date in the form of 'YYYY-MM-DD'. valid_from starts a new version

code

expression Code to be evaluated after populating the temporary table on the database of a time series that is valid from the specified date.

collection_name

character name of a collection to read. Collection are bookmark lists that contain time series keys.

access_level

character describing the access level of the time series or dataset.

set_name

character name of a dataset.

regex

boolean indicating if ts_keys should be interpreted as a regular expression pattern. Defaults to FALSE.

locale

character indicating the language of the meta information to be store. We recommend to use ISO country codes to represent languages. Defaults to NULL. When local is set to NULL, metadata are stored without localization. Note that, when localizing meta information by assigning a language, multiple meta information objects can be stored for a single time series.

respect_release_date

boolean indicating if it should the release embargo of a time series be respected. Defaults to FALSE. This option makes sense when the function is used in an API. In that sense, users do not have direct access to this function and therefore cannot simply switch parameters.

chunksize

set a limit of the number of time series requested in the function.

collection_owner

character username that is the owner of a collection.

user

character name of the database user. Defaults to the user of the R session. this is often the user for the database, too so you do not have to specify your username explicitly if that is the case.